IO classes in a separate io folder
