+++ src/core/io/RamsnapWriter.cpp
