+++ src/core/io/RamtrjRecorder.cpp
