+++ src/core/io/RamtrjRecorder.h
@@ -8,9 +8,9 @@
 #include <iostream>
 #include <memory>
 
-#include "Packing.h"
+#include "core/Packing.h"
 #include "RamtrjIO.h"
-#include "SimulationRecorder.h"
+#include "core/SimulationRecorder.h"
 
 
 /**
