+++ src/core/io/WolframWriter.cpp
