+++ src/core/io/XYZRecorder.cpp
