+++ src/core/io/XYZWriter.cpp
